requirements.txt
# Core
pandas>=2.2
numpy>=1.26
pyarrow>=17.0
duckdb>=1.0

# Embeddings + ML
faiss-cpu>=1.8
scikit-learn>=1.5
sentence-transformers>=3.0

# App + utils
streamlit>=1.37
pydantic>=2.8
pyyaml>=6.0
typer>=0.12
tqdm>=4.66
rich>=13.7
Pillow>=10.4
regex>=2024.5.15

# Optional (uncomment if you want CLIP later; install torch separately per platform)
# open-clip-torch>=2.24
